enabled lint, refactored code, updated workflow
