+++ .github/workflows/webpack.yml
@@ -27,3 +27,4 @@ jobs:
       run: |
         npm install
         npm run compile
+        npm run lint
