--- src/core/handler/impl/checkout_process_handler.ts
@@ -1,6 +1,0 @@
-import { ProcessHandler } from '../process_handler';
-import { AbstractProcessHandler } from './abstract_process_handler';
-
-export class CheckoutProcessHandler extends AbstractProcessHandler implements ProcessHandler {
-
-}
