+++ src/core/handler/impl/delete-process-handler.ts
@@ -1,0 +1,6 @@
+import { ProcessHandler } from '../process-handler';
+import { AbstractProcessHandler } from './abstract-process-handler';
+
+export class DeleteProcessHandler extends AbstractProcessHandler implements ProcessHandler {
+
+}
