+++ src/core/handler/impl/undo-process-handler.ts
@@ -1,0 +1,6 @@
+import { ProcessHandler } from '../process-handler';
+import { AbstractProcessHandler } from './abstract-process-handler';
+
+export class UndoProcessHandler extends AbstractProcessHandler implements ProcessHandler {
+
+}
