--- src/core/handler/impl/undo_process_handler.ts
@@ -1,6 +1,0 @@
-import { ProcessHandler } from '../process_handler';
-import { AbstractProcessHandler } from './abstract_process_handler';
-
-export class UndoProcessHandler extends AbstractProcessHandler implements ProcessHandler {
-
-}
