+++ src/core/process.ts
@@ -1,18 +1,18 @@
-import * as cp from 'child_process';
-import { OutputChannel } from './output_channel';
+import { spawn, spawnSync, ChildProcess } from 'child_process';
+import { OutputChannel } from './output-channel';
 
 export class Process {
-    private childProcess!: cp.ChildProcess;
+    private childProcess!: ChildProcess;
     private commandName!: string;
 
     public spawn(executablePath: string, args: string[]): void {
         this.commandName = args[0]; // The first arg is always a command
-        this.childProcess = cp.spawn(executablePath, args);
+        this.childProcess = spawn(executablePath, args);
     }
 
     public spawnSync(executablePath: string, args: string[]): string {
-        this.commandName = args[0]; // The first arg is always a command        
-        const result = cp.spawnSync(executablePath, args);
+        this.commandName = args[0]; // The first arg is always a command
+        const result = spawnSync(executablePath, args);
         const stdOut = result.stdout;
         OutputChannel.log(stdOut);
         const info = `TF command '${this.getCommandName()}' is successfully done`;
@@ -26,7 +26,10 @@ export class Process {
     }
 
     public write(input: string): void {
-        input && this.childProcess.stdin!.write(input);
+        if (!input) {
+            return;
+        }
+        this.childProcess.stdin!.write(input);
     }
 
     public getCommandName(): string {
