+++ src/core/tfs/impl/delete-tfs-command.ts
@@ -1,0 +1,12 @@
+import { ProcessHandler } from '../../handler/process-handler';
+import { DeleteProcessHandler } from '../../handler/impl/delete-process-handler';
+import { TfsCommandBase } from './tfs-command-base';
+
+export class DeleteTfsCommand extends TfsCommandBase {
+
+    public override readonly command = 'delete';
+
+    public override getConsoleDataHandler(): ProcessHandler {
+        return new DeleteProcessHandler();
+    }
+}
