--- src/core/tfs/impl/undo_tfs_command.ts
@@ -1,12 +1,0 @@
-import { ProcessHandler } from '../../handler/process_handler';
-import { UndoProcessHandler } from '../../handler/impl/undo_process_handler';
-import { TfsCommandBase } from './tfs_command_base';
-
-export class UndoTfsCommand extends TfsCommandBase {
-    
-    public override readonly command = 'undo';
-
-    public override getConsoleDataHandler(): ProcessHandler {
-        return new UndoProcessHandler();
-    }
-}
