+++ src/core/ui/message.ts
@@ -1,15 +1,15 @@
-import * as vscode from 'vscode';
+import { window } from 'vscode';
 
 export class Message {
     public info(message: string, ...items: string[]): Thenable<string | undefined> {
-        return vscode.window.showInformationMessage(message, ...items);
+        return window.showInformationMessage(message, ...items);
     }
 
     public warning(message: string, ...items: string[]): Thenable<string | undefined> {
-        return vscode.window.showWarningMessage(message, ...items);
+        return window.showWarningMessage(message, ...items);
     }
 
     public error(message: string, ...items: string[]): Thenable<string | undefined> {
-        return vscode.window.showErrorMessage(message, ...items);
+        return window.showErrorMessage(message, ...items);
     }
 }
