+++ src/tsconfig.json
@@ -1,0 +1,31 @@
+{
+    "compilerOptions": {
+      "declaration": false,
+      "downlevelIteration": false,
+      "emitDecoratorMetadata": true,
+      "experimentalDecorators": true,
+      "importHelpers": true,
+      "lib": [ "ES2018", "dom" ],
+      "mapRoot": "./",
+      "module": "es2020",
+      "skipLibCheck": true,
+      "moduleResolution": "node",
+      "outDir": "../dist/out-tsc",
+      "sourceMap": true,
+      "target": "es6",
+      "typeRoots": [
+        "../node_modules/@types"
+      ],
+      "baseUrl": "."
+    },
+    "angularCompilerOptions": {
+      "fullTemplateTypeCheck": false,
+      "strictInjectionParameters": false,
+      "enableIvy": false
+    },
+    "include": [
+      "../src/**/*"
+    ],
+    "exclude": [ "node_modules", "**/*.spec.ts" ]
+  }
+
