+++ tslint.json
@@ -1,0 +1,115 @@
+{
+    "rulesDirectory": [
+      "node_modules/codelyzer"
+    ],
+    "rules": {
+      "callable-types": true,
+      "class-name": true,
+      "comment-format": [
+        false,
+        "check-space"
+      ],
+      "curly": true,
+      "eofline": false,
+      "forin": true,
+      "import-blacklist": [ false, "rxjs" ],
+      "import-spacing": true,
+      "indent": [
+        true,
+        "spaces"
+      ],
+      "interface-over-type-literal": true,
+      "label-position": true,
+      "max-line-length": [
+        true,
+        150
+      ],
+      "member-access": false,
+      "member-ordering": [
+        true,
+        {
+          "order": [
+            "static-field",
+            "instance-field",
+            "static-method",
+            "instance-method"
+          ]
+        }
+      ],
+      "no-arg": true,
+      "no-bitwise": true,
+      "no-console": [
+        true,
+        "debug",
+        "info",
+        "time",
+        "timeEnd",
+        "trace"
+      ],
+      "no-construct": true,
+      "no-debugger": true,
+      "no-duplicate-variable": true,
+      "no-empty": false,
+      "no-empty-interface": true,
+      "no-eval": true,
+      "no-inferrable-types": [ true, "ignore-params" ],
+      "no-shadowed-variable": true,
+      "no-string-literal": false,
+      "no-string-throw": false,
+      "no-switch-case-fall-through": true,
+      "no-trailing-whitespace": true,
+      "no-unused-expression": true,    
+      "no-var-keyword": true,
+      "object-literal-sort-keys": false,
+      "one-line": [
+        true,
+        "check-open-brace",
+        "check-catch",
+        "check-else",
+        "check-whitespace"
+      ],
+      "prefer-const": true,
+      "quotemark": [
+        true,
+        "single"
+      ],
+      "radix": true,
+      "semicolon": [true, "always"],
+      "triple-equals": [
+        true,
+        "allow-null-check"
+      ],
+      "typedef-whitespace": [
+        true,
+        {
+          "call-signature": "nospace",
+          "index-signature": "nospace",
+          "parameter": "nospace",
+          "property-declaration": "nospace",
+          "variable-declaration": "nospace"
+        }
+      ],    
+      "unified-signatures": true,
+      "variable-name": false,
+      "whitespace": [
+        true,
+        "check-branch",
+        "check-decl",
+        "check-operator",
+        "check-separator",
+        "check-type"
+      ],
+      "directive-selector": [ true, "attribute", "tc", "camelCase" ],
+      "component-selector": [ true, "element", "tc", "kebab-case" ],
+      "no-inputs-metadata-property": true,
+      "no-outputs-metadata-property": true,
+      "no-host-metadata-property": true,
+      "no-input-rename": false,
+      "no-output-rename": true,
+      "use-lifecycle-interface": true,
+      "use-pipe-transform-interface": true,
+      "component-class-suffix": true,
+      "directive-class-suffix": true     
+    }
+  }
+
